perf(web): lazy-load heavy deps to improve startup
Defer MathJax, Prettier, Mermaid, dialogs, and upload SDKs from the initial bundle. Split web utils modules, load MathJax on demand with safe katex fallbacks, and make cloud sync non-blocking on boot.
