+++ apps/web/index.html
@@ -9,7 +9,7 @@
       name="viewport"
       content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, viewport-fit=cover"
     />
-    <title>微信 Markdown 编辑器 | Doocs 开源社区</title>
+    <title>微信 Markdown 编辑器 | Doocs</title>
     <link rel="shortcut icon" href="/favicon.png" />
     <link rel="apple-touch-icon-precomposed" href="/favicon.png" />
     <style>
@@ -92,16 +92,6 @@
         }
       }, 30000)
     </script>
-    <script>
-      window.MathJax = {
-        tex: { tags: 'ams' },
-        svg: { fontCache: 'none' },
-      }
-    </script>
-    <script
-      id="MathJax-script"
-      src="https://cdn-doocs.oss-cn-shenzhen.aliyuncs.com/npm/mathjax@3/es5/tex-svg.js"
-    ></script>
     <script type="module" src="/src/main.ts"></script>
     <script type="module" src="/src/sidepanel.ts"></script>
   </body>
