+++ apps/web/src/components/editor/dialogs/InsertFormDialog.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { useEditorStore } from '@/stores/editor'
 import { useUIStore } from '@/stores/ui'
-import { createTable } from '@/utils'
+import { createTable } from '@/utils/shared-helpers'
 
 const editorStore = useEditorStore()
 const uiStore = useUIStore()
