+++ apps/web/src/main.ts
@@ -1,4 +1,4 @@
-import { initComponentDarkVars, initializeMermaid } from '@md/core'
+import { initComponentDarkVars } from '@md/core'
 import { createPinia } from 'pinia'
 import { createApp } from 'vue'
 import App from './App.vue'
@@ -11,8 +11,6 @@ import 'vue-sonner/style.css'
 import '@/assets/index.css'
 import '@/assets/less/theme.less'
 
-// 异步初始化 mermaid，避免初始化顺序问题
-initializeMermaid().catch(console.error)
 // 注入组件暗色模式 CSS 变量（独立样式元素，不污染剪贴板）
 initComponentDarkVars()
