+++ apps/web/src/stores/mpAccounts.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid'
-import { addPrefix } from '@/utils'
+import { addPrefix } from '@/utils/prefix'
 import { store } from '@/utils/storage'
 
 export interface MpAccount {
