+++ apps/web/src/utils/file-upload.ts
@@ -1,0 +1,4 @@
+export async function fileUpload(content: string, file: File): Promise<string> {
+  const { fileUpload: upload } = await import(`./file`)
+  return upload(content, file)
+}
