+++ apps/web/src/utils/mathjax.ts
@@ -1,0 +1,9 @@
+export {
+  contentHasMath,
+  ensureMathJaxLoaded,
+  isMathJaxReady,
+  loadMathJax,
+  MATHJAX_READY_EVENT,
+} from '@md/core/utils'
+
+export { isMathJaxReady as isMathJaxLoaded } from '@md/core/utils'
