+++ apps/web/src/utils/prefix.ts
@@ -1,0 +1,5 @@
+import { prefix } from '@md/shared/configs'
+
+export function addPrefix(str: string) {
+  return `${prefix}__${str}`
+}
