+++ packages/core/src/utils/index.ts
@@ -2,4 +2,6 @@ export { initializeMermaid } from '../extensions/mermaid'
 export * from './basicHelpers'
 export * from './languages'
 export * from './markdownHelpers'
+export * from './mathDetection'
+export * from './mathjax'
 export * from './svgCache'
