+++ packages/shared/src/editor/javascript.ts
@@ -1,7 +1,6 @@
 import type { EditorView } from '@codemirror/view'
 import { javascript } from '@codemirror/lang-javascript'
 import { keymap } from '@codemirror/view'
-import { formatDoc } from '../utils/fileHelpers'
 import { basicSetup } from './basicSetup'
 
 /**
@@ -9,6 +8,7 @@ import { basicSetup } from './basicSetup'
  */
 async function formatJavaScript(view: EditorView) {
   const content = view.state.doc.toString()
+  const { formatDoc } = await import('../utils/formatDoc')
   const formatted = await formatDoc(content, `javascript`)
   view.dispatch({
     changes: { from: 0, to: view.state.doc.length, insert: formatted },
